pyproject.toml
[project]
name = "soop-autorec"
version = "0.1.0"
description = "SOOP channel auto recording service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cryptography>=44.0.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.4",
  "pydantic-settings>=2.7.0",
  "python-multipart>=0.0.18",
  "streamlink>=8.2.1",
  "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".uv-cache", ".venv", "__pycache__", "data"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
